Updating branch Fernando - 4/22/2020

+++ Database/DatabaseDelete.java
@@ -3,61 +3,79 @@ package Database;
 /*
  * This is a Database class for deleting records.
  *
- * Last Updated: April 19, 2020
+ * Last Updated: April 20, 2020
  * @author Will Higdon, Quinn Tjin-A-Soe, Fernando Villarreal
  */
+import Objects.UserObject;
+import java.io.BufferedReader;
 import Objects.UserObject;
 import Objects.RecordObject;
 import java.io.BufferedWriter;
 import java.io.File;
+import java.io.FileNotFoundException;
+import java.io.FileReader;
 import java.io.FileWriter;
 import java.io.IOException;
-import java.util.ArrayList;
 import java.util.Scanner;
 
 public class DatabaseDelete {
+    public static final String active = "true";
+    public static final String inactive = "false";
 
     //==================== PUBLIC METHODS ====================
-
     /**
      * This method deletes a specified record from the database. This method
-     * scans the entire file for the specified record. If it exists, then it
-     * will set it to "inactive" and overwrite the previous file.
+     * writes the contents of one file onto another text file except for the
+     * specified record, which will be set to inactive onto the temporary text
+     * file. Afterwards, the temporary text file is renamed to the original text
+     * file, and the original text file is deleted.
      *
      * @param _file
      * @param _user
+     * @throws java.io.FileNotFoundException
      * @throws IOException
      */
-    public static void deleteUserRecord(File _file, UserObject _user) throws IOException, Exception {
-        Scanner scanner = new Scanner(_file);
+    public static void deleteUserRecord(File _file, UserObject _user) throws FileNotFoundException, IOException, Exception {
+        File originalFile = _file;
+        BufferedReader reader = new BufferedReader(new FileReader(originalFile));
+        BufferedWriter writer = new BufferedWriter(new FileWriter(DatabaseInterface.temporaryFile));
+        String lineInFile = null;
         String userName = _user.getUserName();
-        ArrayList<String> arrayRecord = new ArrayList<>();
-        String stringRecord = "";
-        scanner.useDelimiter("\t");
-
-        while (scanner.hasNextLine()) {
-            String nextLine = scanner.nextLine();
-            if (nextLine.contains(userName) && nextLine.contains("" + DatabaseInterface.active)) {
-                stringRecord += _user.getUuid() + "\t";
-                stringRecord += _user.getUserName() + "\t";
-                stringRecord += _user.getUserPassword() + "\t";
-                stringRecord += _user.getUserEmail() + "\t";
-                stringRecord += _user.getUserFirstName() + "\t";
-                stringRecord += _user.getUserLastName() + "\t";
-                stringRecord += DatabaseInterface.inactive + "\n";
+        // While the file is not empty
+        while ((lineInFile = reader.readLine()) != null) {
+            // If the username is found, it will be made to inactive
+            if (lineInFile.contains(userName)) {
+                lineInFile = lineInFile.replace(DatabaseDelete.active, DatabaseDelete.inactive);
+                writer.write(lineInFile);
+                writer.newLine();
             } else {
-                stringRecord += nextLine + "\n";
+                writer.write(lineInFile);
+                writer.newLine();
             }
+            writer.flush();
         }
-        BufferedWriter writer = new BufferedWriter(new FileWriter(_file, false));
-        writer.write(stringRecord);
         writer.close();
+        reader.close();
+
+        // Delete the original file
+        if (!originalFile.delete()) {
+            System.out.println("Could not delete file");
+            return;
+        }
+
+        // Rename the new file to the filename the original file had.
+        if (!DatabaseInterface.temporaryFile.renameTo(originalFile)) {
+            System.out.println("Could not rename file");
+        }
+
         // Delete the user's information from the UserLogin.txt file
         DatabaseDelete.deleteUserLogin(_user);
+
     }
 
     /**
      * Deletes the given user from the database.
+     *
      * @param _user
      * @throws Exception
      */
@@ -67,7 +85,9 @@ public class DatabaseDelete {
 
     /**
      * Deletes the user associated with the given username from the database.
+     *
      * @param _username
+     * @throws java.io.IOException
      * @throws Exception
      */
     public static void deleteUserRecord(String _username) throws IOException, Exception {
@@ -78,9 +98,9 @@ public class DatabaseDelete {
     }
 
     //==================== PRIVATE METHODS ====================
-
     /**
      * Deletes a user's login information from the UserLogin.txt file.
+     *
      * @param _user
      * @throws Exception
      */
@@ -97,7 +117,7 @@ public class DatabaseDelete {
             if (nextLine.contains(username) && nextLine.contains("" + DatabaseInterface.active)) {
                 recordsString += username + "\t" + _user.getUserPassword() + "\t";
                 recordsString += DatabaseInterface.inactive + "\n";
-            // If the record does not contain the username, append it to recordsString without overwriting.
+                // If the record does not contain the username, append it to recordsString without overwriting.
             } else {
                 recordsString += nextLine + "\n";
             }
